Added missing server structure
