+++ include/Channel.hpp
@@ -1,0 +1,27 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   Channel.hpp                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: gilmar <gilmar@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/05/21 08:24:04 by gilmar            #+#    #+#             */
+/*   Updated: 2024/05/21 08:28:53 by gilmar           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include <string>
+#include <vector>
+
+class Channel
+{
+    public:
+        Channel();
+        ~Channel();
+    private:
+        int topic;
+        std::string name;
+        std::string password;
+        std::string created_at;
+        // TODO: Add a vector of clients
+};
