+++ src/Channel.cpp
@@ -1,0 +1,25 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   Channel.cpp                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: gilmar <gilmar@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/05/21 08:26:17 by gilmar            #+#    #+#             */
+/*   Updated: 2024/05/21 08:26:28 by gilmar           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "Channel.hpp"
+
+Channel::Channel()
+{
+    topic = 0;
+    name = "";
+    password = "";
+    created_at = "";
+}
+
+Channel::~Channel()
+{
+}
