+++ src/Server.cpp
@@ -6,7 +6,7 @@
 /*   By: gilmar <gilmar@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/28 10:26:55 by gilmar            #+#    #+#             */
-/*   Updated: 2024/05/19 19:33:36 by gilmar           ###   ########.fr       */
+/*   Updated: 2024/05/21 08:54:05 by gilmar           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -234,7 +234,8 @@ void Server::_execute_command(const std::string buffer, const int fd)
         return ;
     std::string command = toupper(splitted_buffer[0]);
     std::string parameters = splitted_buffer[1];
-    
+
+    //TODO: Add a default handler for unknown commands
     for (size_t i = 0; i < _command_list_size; i++) {
         if (command == _command_list[i].command) {
             (this->*_command_list[i].handler)(parameters, fd);
@@ -243,272 +244,7 @@ void Server::_execute_command(const std::string buffer, const int fd)
     }
 }
 
-/*
-** ------------------------------- JOIN COMMAND --------------------------------
-*/
 
-/**
- * @brief Handles the JOIN command received from the client.
- *
- * This method processes the JOIN command received from the client and sends a response
- * to the client indicating that the client has joined the channel.
- *
- * @param buffer The buffer containing the JOIN command parameters.
- * @param fd The file descriptor associated with the client that sent the command.
- */
-void Server::_handler_client_join(const std::string &buffer, const int fd)
-{
-    Client* client = _get_client(fd);
-
-    std::string channel = buffer;
-    
-    _send_response(fd, RPL_TOPICIS(client->get_nickname(), channel, std::string("Welcome to the channel")));
-}
-
-/*
-** ------------------------------- QUIT COMMAND --------------------------------
-*/
-
-/**
- * @brief Handles the QUIT command received from the client.
- *
- * This method processes the QUIT command received from the client and sends a response
- * to the client indicating that the client has quit the server.
- *
- * @param buffer The buffer containing the QUIT command parameters.
- * @param fd The file descriptor associated with the client that sent the command.
- */
-void Server::_handler_client_quit(const std::string &buffer, const int fd)
-{
-    (void)buffer;
-    _send_response(fd, RPL_QUITMESSAGE(_get_client(fd)->get_nickname()));
-    _clear_client(fd);
-}
-
-/*
-** ------------------------------- PART COMMAND --------------------------------
-*/
-
-/**
- * @brief Handles the PART command received from the client.
- *
- * This method processes the PART command received from the client and sends a response
- * to the client indicating that the client has left the channel.
- *
- * @param buffer The buffer containing the PART command parameters.
- * @param fd The file descriptor associated with the client that sent the command.
- */
-void Server::_handler_client_part(const std::string &buffer, const int fd)
-{
-    (void)fd;
-    (void)buffer;
-    Client* client = _get_client(fd);
-
-    _send_response(fd, RPL_PART(client->get_hostname(), "ft_trancendence", client->get_nickname()));
-}
-
-/*
-** ------------------------------- MODE COMMAND --------------------------------
-*/
-
-/**
- * @brief Handles the MODE command received from the client.
- *
- * This method processes the MODE command received from the client and sends a response
- * to the client indicating the channel modes supported by the server.
- *
- * @param buffer The buffer containing the MODE command parameters.
- * @param fd The file descriptor associated with the client that sent the command.
- */
-void Server::_handler_client_mode(const std::string &buffer, const int fd)
-{
-    (void)fd;
-    (void)buffer;
-    Client* client = _get_client(fd);
-
-    _send_response(fd, RPL_CHANNELMODES(client->get_nickname(), "ft_trancendence", "+nt"));
-}
-
-/*
-** ------------------------------- KICK COMMAND --------------------------------
-*/
-
-/**
- * @brief Handles the KICK command received from the client.
- *
- * This method processes the KICK command received from the client and sends a response
- * to the client indicating that the client has been kicked from the channel.
- *
- * @param buffer The buffer containing the KICK command parameters.
- * @param fd The file descriptor associated with the client that sent the command.
- */
-void Server::_handler_client_kick(const std::string &buffer, const int fd)
-{
-    (void)fd;
-    (void)buffer;
-    Client* client = _get_client(fd);
-
-    _send_response(fd, RPL_KICK(client->get_nickname(), "ft_trancendence", "gilmar", "Ygor", "You have been kicked"));
-}
-
-/*
-** ------------------------------- TOPIC COMMAND --------------------------------
-*/
-
-/**
- * @brief Handles the TOPIC command received from the client.
- *
- * This method processes the TOPIC command received from the client and sends a response
- * to the client indicating the topic of the channel.
- *
- * @param buffer The buffer containing the TOPIC command parameters.
- * @param fd The file descriptor associated with the client that sent the command.
- */
-void Server::_handler_client_topic(const std::string &buffer, const int fd)
-{
-    (void)fd;
-    (void)buffer;
-    Client* client = _get_client(fd);
-
-    _send_response(fd, RPL_TOPICIS(client->get_nickname(), "ft_trancendence", "Welcome to the channel"));
-}
-
-/*
-** ------------------------------- INVITE COMMAND --------------------------------
-*/
-
-/**
- * @brief Handles the INVITE command received from the client.
- *
- * This method processes the INVITE command received from the client and sends a response
- * to the client indicating that the client has been invited to the channel.
- *
- * @param buffer The buffer containing the INVITE command parameters.
- * @param fd The file descriptor associated with the client that sent the command.
- */
-void Server::_handler_client_invite(const std::string &buffer, const int fd)
-{
-    (void)fd;
-    (void)buffer;
-    Client* client = _get_client(fd);
-
-    _send_response(fd, RPL_INVITING(client->get_hostname(), "ft_trancendence", "Gilmar", "Ygor"));
-}
-
-/*
-** ------------------------------- INVITE COMMAND --------------------------------
-*/
-
-/**
- * @brief Handles the PRIVMSG command received from the client.
- *
- * This method processes the PRIVMSG command received from the client and sends a response
- * to the client indicating that the message has been sent.
- *
- * @param buffer The buffer containing the PRIVMSG command parameters.
- * @param fd The file descriptor associated with the client that sent the command.
- */
-void Server::_handler_client_privmsg(const std::string &buffer, const int fd)
-{
-    (void)fd;
-    (void)buffer;
-    Client* client = _get_client(fd);
-
-    _send_response(fd, RPL_PRIVMSG(client->get_hostname(), "ft_trancendence", "Hello, Carlos!"));
-}
-
-/*
-** ------------------------------- NICK COMMAND --------------------------------
-*/
-
-/**
- * @brief Handles the NICK command received from the client.
- *
- * This method processes the NICK command received from the client and sends a response
- * to the client indicating that the client's nickname has been changed.
- *
- * @param nickname The new nickname to be assigned to the client.
- * @param fd The file descriptor associated with the client that sent the command.
- */
-void Server::_handler_client_nickname(const std::string &nickname, const int fd)
-{
-    Client* client = _get_client(fd);
-
-    if (nickname.empty() || nickname.size() < 5)
-        _send_response(fd, ERR_NOTENOUGHPARAM(std::string("*")));
-    else if (!_is_valid_nickname(nickname))
-        _send_response(fd, ERR_ERRONEUSNICK(client->get_nickname()));
-    else if (_is_nickname_in_use(fd, nickname))
-        _send_response(fd, ERR_NICKINUSE(client->get_nickname()));
-    else if (!client->get_is_registered())
-        _send_response(fd, ERR_NOTREGISTERED(std::string("*")));
-    else
-    {
-        client->set_nickname(nickname);
-        if (_client_is_ready_to_login(fd))
-            client->set_is_logged(fd);
-    }
-}
-
-/*
-** ------------------------------- USERNAME COMMAND --------------------------------
-*/
-
-/**
- * @brief Handles the USERNAME command received from the client.
- *
- * This method processes the USERNAME command received from the client and sends a response
- * to the client indicating that the client's username has been set.
- *
- * @param username The username to be assigned to the client.
- * @param fd The file descriptor associated with the client that sent the command.
- */
-void Server::_handler_client_username(const std::string &username, const int fd)
-{
-    Client* client = _get_client(fd);
-
-    if (username.empty() || username.size() < 5)
-        _send_response(fd, ERR_NOTENOUGHPARAM(std::string("*")));
-    else if (!client->get_is_registered())
-        _send_response(fd, ERR_NOTREGISTERED(std::string("*")));
-    else if (!client->get_username().empty())
-        _send_response(fd, ERR_ALREADYREGISTERED(std::string("*")));
-    else
-    {
-        client->set_username(username);
-        if (_client_is_ready_to_login(fd))
-            client->set_is_logged(fd);
-    }
-}
-
-/*
-** ------------------------------- PASSWORD COMMAND --------------------------------
-*/
-
-/**
- * @brief Handles the PASSWORD command received from the client.
- *
- * This method processes the PASSWORD command received from the client and sends a response
- * to the client indicating that the client's password has been set.
- *
- * @param password The password to be assigned to the client.
- * @param fd The file descriptor associated with the client that sent the command.
- */
-void Server::_handler_client_password(const std::string &password, const int fd)
-{
-    Client* client = _get_client(fd);
-
-    if (password.empty()) //-> Isso é o suficiente?
-        _send_response(fd, ERR_NOTENOUGHPARAM(std::string("*")));
-    else if (client->get_is_registered())
-        _send_response(fd, ERR_ALREADYREGISTERED(std::string("*")));
-    else if (_password != password)
-        _send_response(fd, ERR_INCORPASS(std::string("*")));
-    else
-    {
-        client->set_is_registered(true);
-    }
-}
 
 /*
 ** ------------------------------- PARSER COMMAND --------------------------------
