+++ src/commands/Invite.cpp
@@ -1,0 +1,37 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   Invite.cpp                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: gilmar <gilmar@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/05/21 08:31:16 by gilmar            #+#    #+#             */
+/*   Updated: 2024/05/21 08:55:05 by gilmar           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "Server.hpp"
+
+/*
+ * Command: INVITE
+ * Parameters: <nickname> <channel>
+ * Link: https://datatracker.ietf.org/doc/html/rfc1459#section-4.2.7
+*/
+
+/**
+ * @brief Handles the INVITE command received from the client.
+ *
+ * This method processes the INVITE command received from the client and sends a response
+ * to the client indicating that the client has been invited to the channel.
+ *
+ * @param buffer The buffer containing the INVITE command parameters.
+ * @param fd The file descriptor associated with the client that sent the command.
+ */
+void Server::_handler_client_invite(const std::string &buffer, const int fd)
+{
+    (void)fd;
+    (void)buffer;
+    Client* client = _get_client(fd);
+
+    _send_response(fd, RPL_INVITING(client->get_hostname(), "ft_trancendence", "Gilmar", "Ygor"));
+}
